requirements.txt
flask
pyodbc
python-dotenv
gunicorn
pylint